requirements.txt
fastapi
uvicorn
sqlalchemy
pymssql
pyodbc
python-dotenv
pydantic
httpx